Keyboard navigation fixed for non-search screens
