+++ ui/src/components/common/Entity.jsx
@@ -67,13 +67,33 @@ class EntityLabel extends Component {
 }
 
 class EntityLink extends Component {
+   static openPreview({ history, location, entity}){
+    const parsedHash = queryString.parse(location.hash);
+    const previewType = entity.schemata.indexOf('Document') !== -1 ? 'document' : 'entity';
+
+    if (parsedHash['preview:id'] === entity.id && parsedHash['preview:type'] === previewType) {
+      parsedHash['preview:id'] = undefined;
+      parsedHash['preview:type'] = undefined;
+      parsedHash['preview:mode'] = undefined;
+    } else {
+      parsedHash['preview:id'] = entity.id;
+      parsedHash['preview:type'] = previewType;
+      parsedHash['preview:mode'] = undefined;
+    }
+    history.replace({
+      pathname: location.pathname,
+      search: location.search,
+      hash: queryString.stringify(parsedHash),
+    });
+
+  }
   onClick = (event) => {
-    const { entity, history,preview } = this.props;
+    const { entity, history, preview } = this.props;
 
     event.preventDefault();
 
     if (preview) {
-      preview(entity);
+      Entity.constructor.openPreview(this.props)
     } else {
       history.push({
         pathname: getPath(entity.links.ui)
